renamed colorlist component, general css fixes
