+++ resources/css/app.scss
@@ -122,6 +122,10 @@ input::-webkit-inner-spin-button {
     padding-right: 0;
 }
 
+.noPaddingRight {
+    padding-right: 0;
+}
+
 .col-lg-2-5 {
     flex: 0 0 20.83333333335%;
     max-width: 20.83333333335%;
@@ -578,7 +582,7 @@ li.color:nth-child(16) { background: #6b6e53; } li.color:nth-child(17) { backgro
 .front,
 .back {
     width: 100%;
-    height: 310px;
+    height: 276px;
 }
 
 .flip-card {
@@ -608,7 +612,7 @@ li.color:nth-child(16) { background: #6b6e53; } li.color:nth-child(17) { backgro
 .motivPaket {
     position: relative;
     box-sizing: border-box;
-    height: 280px;
+    height: 250px;
     width: 100%;
     overflow: hidden;
     margin: 10px;
@@ -1507,7 +1511,7 @@ hr {
     .card-container,
     .front,
     .back {
-        height: 250px;
+        height: 226px;
         -ms-user-select: None;
         -moz-user-select: None;
         -webkit-user-select: None;
@@ -1515,7 +1519,7 @@ hr {
     }
 
     .motivPaket {
-        height: 220px;
+        height: 200px;
     }
 
     .motivVorschauBild canvas {
@@ -1530,9 +1534,17 @@ hr {
 }
 
 @media (max-width: 991px) {
+    .noPaddingRight {
+        padding-right: 15px;
+    }
+
     #msform .stepButton {
         font-size: 10px;
         width: 188px;
+    }
+
+    .cardProdukt { 
+        margin-bottom: 30px;
     }
 
     .produktPreis {
@@ -1604,10 +1616,6 @@ hr {
 
     #msform fieldset .form-card {
         padding: 20px 5px;
-    }
-
-    .cardProdukt { 
-        margin-bottom: 30px;
     }
 
     .product-cards > div:last-child {
