+++ resources/js/Components/ColorList.vue
